refactor(backend): extract DatabaseCore from OpenBotDatabase
Moves the SQLite handle, its lifecycle and migration, and the append-events /
project / write-receipt primitive into `src/backend/database/database-core.ts`.
The facade keeps its class identity, constructor signature and full public
surface, and now holds a `#core` it delegates to — including `get connection()`
and `dispatch()`, both of which external modules use directly.

The body of `dispatch` and of `deleteEventsAndReceipt` is moved verbatim: both
independently repeat `const ownsTransaction = !db.isTransaction`, which is what
lets a projector nest another dispatch inside the caller's transaction. Neither
copy is harmonized.

Controllers read `core.connection` at each use rather than caching it: a cached
handle survives `initialize()` and then silently addresses a closed database
after `close()` + reopen.
